Commit wspr tone frequency patch and fix GrayLineDuration typo
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5465 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ Modulator.cpp
@@ -43,7 +43,7 @@ Modulator::Modulator (unsigned frameRate, unsigned periodLengthInSeconds,
 }
 
 void Modulator::start (unsigned symbolsLength, double framesPerSymbol,
-                       unsigned frequency, double toneSpacing,
+                       double frequency, double toneSpacing,
                        SoundOutput * stream, Channel channel,
                        bool synchronize, double dBSNR)
 {
