+++ mainwindow.h
@@ -246,11 +246,11 @@ private:
   Q_SIGNAL void startDetector (AudioDevice::Channel) const;
   Q_SIGNAL void detectorClose () const;
   Q_SIGNAL void finished () const;
-  Q_SIGNAL void transmitFrequency (unsigned) const;
+  Q_SIGNAL void transmitFrequency (double) const;
   Q_SIGNAL void endTransmitMessage (bool quick = false) const;
   Q_SIGNAL void tune (bool = true) const;
   Q_SIGNAL void sendMessage (unsigned symbolsLength, double framesPerSymbol,
-      unsigned frequency, double toneSpacing,
+      double frequency, double toneSpacing,
       SoundOutput *, AudioDevice::Channel = AudioDevice::Mono,
       bool synchronize = true, double dBSNR = 99.) const;
   Q_SIGNAL void outAttenuationChanged (qreal) const;
